requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database (psycopg3 - compatible with Python 3.13)
psycopg[binary]==3.2.3

# Cache & Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# API Documentation
drf-spectacular==0.26.5

# CORS
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# Image Processing
Pillow==11.0.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

# Payment Integration (Paystack)
pypaystack2==2.0.0

# WhatsApp Integration & HTTP
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Security (Password Hashing - Argon2)
argon2-cffi==23.1.0
